+++ platform/CodatPlatform/Models/Shared/SyncSetting.cs
@@ -8,7 +8,7 @@
 // </auto-generated>
 //------------------------------------------------------------------------------
 #nullable enable
-namespace CodatPlatform.Models.Shared
+namespace Codat.Platform.Models.Shared
 {
     using Newtonsoft.Json;
     
@@ -22,7 +22,7 @@ namespace CodatPlatform.Models.Shared
         /// Available Data types
         /// </summary>
         [JsonProperty("dataType")]
-        public DataType DataType { get; set; } = default!;
+        public Models.Shared.DataType DataType { get; set; } = default!;
 
         /// <summary>
         /// Whether this data type should be queued after a company has authorized a connection.