+++ platform/RELEASES.md
@@ -77,3 +77,13 @@ Based on:
 - [csharp v2.2.0] platform
 ### Releases
 - [NuGet v2.2.0] https://www.nuget.org/packages/Codat.Platform/2.2.0 - platform
+
+## 2023-11-27 01:25:14
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Platform.yaml
+- Speakeasy CLI 1.122.0 (2.195.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [csharp v3.0.0] platform
+### Releases
+- [NuGet v3.0.0] https://www.nuget.org/packages/Codat.Platform/3.0.0 - platform